refactor(cli): Merge logic + tests from temporary manifest_v2.rs back to manifest.rs
